dotconf_register_options should properly handle failure.
Prior to this commit, if this function failed to allocate or reallocate memory,
it just kept on plugging.
It would crash the program when it accessed some invalid memory address
close to 0.
Now, it detects allocation failures, and it won't access invalid
memory.  Furthermore, if realloc fails, the configfile parameter is left
untouched.

I also changed the return type of this function from void to int.
When it fails, it needs to be able to inform its caller.