remove -release and --export-dynamic flags
The libtool manual states that if you use the -release flag, none of
your releases will be binary compatible with each other, so I have
removed this flag.

Also, this is a library and will not be used with the dlopen call, so I
removed the --export-dynamic flag as well.