fix "modules" example
The modules test was relying on the $USER environment variable as part
of the output.  This insured that this test would fail unless the USER
environment variable was set to lukas.  The fix was to use a different
environment variable, "TESTUSER" which is set to"lukas" inside the
program and retrieved when the config file is processed.