fix include path processing
In dotconf_create, if DC_INCLUDEPATH is not set, set the default for
includepath to be the same directory where the configuration file is
located if possible.  This makes much more sense than using the current
working directory as the default for include files.

If this cannot be done, includepath is left empty by default.